+++ AutoBid_BE/src/test/java/com/codesquad/autobid/car/controller/CarControllerTest.java
@@ -3,10 +3,10 @@ package com.codesquad.autobid.car.controller;
 import com.codesquad.autobid.car.domain.Car;
 import com.codesquad.autobid.car.domain.CheckCarResponse;
 import com.codesquad.autobid.car.repository.CarRepository;
-import com.codesquad.autobid.car.util.CarTestUtil;
+import com.codesquad.autobid.util.CarTestUtil;
 import com.codesquad.autobid.user.domain.User;
 import com.codesquad.autobid.user.repository.UserRepository;
-import com.codesquad.autobid.user.util.UserTestUtil;
+import com.codesquad.autobid.util.UserTestUtil;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
